src/requirements.txt
# wxPython GUI 依赖
wxPython>=4.2.0

# YAML 处理
PyYAML>=6.0

# 环境配置
python-dotenv>=1.0.0

# 网络请求
requests>=2.31.0

# 日志
loguru>=0.7.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 文档处理
python-docx>=1.0.0
markdown>=3.5.0
PyPDF2>=3.0.0

# Web API
flask>=3.0.0
flask-cors>=4.0.0

# 网络检索
playwright>=1.40.0
beautifulsoup4>=4.12.0
httpx>=0.25.0

# 语义检索
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# 数据分析
statsmodels>=0.14.0
matplotlib>=3.8.0
seaborn>=0.13.0

# 进度条
tqdm>=4.66.0
